requirements.txt
libvirt-python>=9.0.0
paramiko>=3.0.0
pyyaml>=6.0

